chore: remove old runtime core
