+++ packages/core/src/elements/UIElement.ts
@@ -1,38 +1,43 @@
-import type { Axis, AxisBox, AxisStyle, AxisConstraints, MeasureCtx } from '../layout/engine.js';
+import type { MeasureCtx } from '../layout/engine.js';
 import { RuleRegistry } from '../layout/engine.js';
+import { createDefaultRegistry } from '../layout/register-defaults.js';
 
-export interface Size { width: number; height: number; }
+export type Size = { width: number; height: number };
+export type Rect = { x: number; y: number; width: number; height: number };
 
-function defaultBox(): AxisBox {
-  return { marginStart: 0, marginEnd: 0, borderStart: 0, borderEnd: 0, paddingStart: 0, paddingEnd: 0 };
-}
+const defaultRegistry = createDefaultRegistry();
 
-export class UIElement {
-  protected boxX: AxisBox = defaultBox();
-  protected boxY: AxisBox = defaultBox();
-  protected styleX: AxisStyle = { unit: 'auto' };
-  protected styleY: AxisStyle = { unit: 'auto' };
-  protected intrinsicWidth?: number;
-  protected intrinsicHeight?: number;
+/**
+ * Basic UI element supporting margin and preferred/min sizes. The size
+ * computation is delegated to the core RuleRegistry to avoid duplicating
+ * clamp logic across runtime elements.
+ */
+export abstract class UIElement {
   desired: Size = { width: 0, height: 0 };
+  final: Rect = { x: 0, y: 0, width: 0, height: 0 };
 
-  constructor(protected registry: RuleRegistry) {}
+  margin = { l: 0, t: 0, r: 0, b: 0 };
+  minW = 0; minH = 0;
+  prefW?: number; prefH?: number;
 
-  protected ctx(axis: Axis, available: number): MeasureCtx {
-    const style = axis === 'x' ? this.styleX : this.styleY;
-    const box = axis === 'x' ? this.boxX : this.boxY;
-    const intrinsic = axis === 'x' ? this.intrinsicWidth : this.intrinsicHeight;
-    const pairedIntrinsic = axis === 'x' ? this.intrinsicHeight : this.intrinsicWidth;
-    const constraints: AxisConstraints = { available };
-    return { axis, box, style, constraints, intrinsic, pairedIntrinsic };
+  protected registry: RuleRegistry = defaultRegistry;
+
+  protected measureAxis(axis: 'x' | 'y', avail: number, intrinsic: number): number {
+    const pref = axis === 'x' ? this.prefW : this.prefH;
+    const marginSum = axis === 'x' ? this.margin.l + this.margin.r : this.margin.t + this.margin.b;
+    const min = (axis === 'x' ? this.minW : this.minH) + marginSum;
+    const prefVal = pref !== undefined ? pref + marginSum : undefined;
+    const style = prefVal !== undefined ? { unit: 'px', value: prefVal } as const : { unit: 'auto' } as const;
+    const ctx: MeasureCtx = {
+      axis,
+      box: { marginStart: 0, marginEnd: 0, borderStart: 0, borderEnd: 0, paddingStart: 0, paddingEnd: 0 },
+      style,
+      constraints: { available: avail, min },
+      intrinsic,
+    };
+    return this.registry.run(ctx);
   }
 
-  protected ctxX(avail: number): MeasureCtx { return this.ctx('x', avail); }
-  protected ctxY(avail: number): MeasureCtx { return this.ctx('y', avail); }
-
-  measure(avail: Size): void {
-    const width = this.registry.run(this.ctxX(avail.width));
-    const height = this.registry.run(this.ctxY(avail.height));
-    this.desired = { width, height };
-  }
+  abstract measure(avail: Size): void;
+  abstract arrange(rect: Rect): void;
 }
