+++ packages/core/src/index.ts
@@ -3,5 +3,3 @@ export * from './layout/register-defaults.js';
 export * from './layout/rules/basic.js';
 export * from './layout/rules/extra.js';
 export * from './elements/UIElement.js';
-export * from './elements/LeafElement.js';
-export * from './elements/SingleChildElement.js';
