+++ packages/parser/package.json
@@ -9,7 +9,8 @@
     "build": "tsc -p tsconfig.json"
   },
   "dependencies": {
-    "pixi.js": "^7.4.3"
+    "pixi.js": "^7.4.3",
+    "@noxigui/core": "file:../core"
   },
   "devDependencies": {
     "typescript": "~5.8.3"
