+++ packages/parser/src/parsers/TextBlockParser.ts
@@ -2,7 +2,7 @@ import { Text } from '../../../runtime/src/elements/Text.js';
 import { applyGridAttachedProps, parseSizeAttrs, applyMargin, applyAlignment } from '../../../runtime/src/helpers.js';
 import type { ElementParser } from './ElementParser.js';
 import type { Parser } from '../Parser.js';
-import type { UIElement } from '../../../runtime/src/core.js';
+import type { UIElement } from '../../../core/src/index.js';
 import type * as PIXI from 'pixi.js';
 
 /** Parser for `<TextBlock>` elements. */
