+++ packages/runtime/src/index.ts
@@ -1,10 +1,10 @@
 export * from './template.js';
-export * from './core.js';
 export { createDefaultRegistry, RuleRegistry } from '../../core/src/index.js';
 export * from './elements/BorderPanel.js';
 export * from './elements/Image.js';
 export * from './elements/Text.js';
 export * from './elements/Grid.js';
+export * from './elements/ContentPresenter.js';
 export * from './helpers.js';
 export type { Renderer, RenderImage, RenderText, RenderGraphics, RenderContainer } from './renderer.js';
 export { RuntimeInstance } from './runtime.js';
